ion: support multiple address space map
Ion doesn't support multiple address space. The recent IOMMU usually
provide multiple address spaces, where a mapping depends on an address
space. An address space is bind to a device pointer passed from
attachment. Those info needs to be stored in a buffer info. If the
same buffer needs to be mapped into the same AS, the same mapping info
needs to be passed back.

Bug 1450489

Change-Id: I0e0b01006b8f167c975652f222b7d465ef899f6d
Signed-off-by: Hiroshi Doyu <hdoyu@nvidia.com>
Reviewed-on: http://git-master/r/377246
Reviewed-by: Automatic_Commit_Validation_User