bcmdhd: workaround for compiling in bcmdhd module
Compile bcmdhd once, currently when building bcmdhd for bcm43241 and
bcm4331 a custom makefile is used. When compiling in bcmdhd, it
causes the driver to be built with the targeted driver makefile
and the makefile of bcmdhd. This causes an error due to multiple
defines as the driver is compiled twice.